backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
pydantic==2.12.5
pydantic-settings==2.7.0
google-genai==1.67.0
google-adk==1.27.1
yfinance==1.2.0
ta==0.11.0
newsapi-python==0.2.7
firebase-admin==6.7.0
python-multipart==0.0.20
